Rename file main-nav.vue to MainNav.vue
